NDRS-763: Extract trait to a separate type.
